split into GrantSelector, SelectGroupModal, and GroupMembersLabel
